Refactor after all encryption tests passing
Added tests for a encrypting with a variable shift.
Added a separate method encrypt which took two variables.
Previous method was renamed encrypt_by_five and only accepted one variable.
All the tests passed.
Refactored to combine the two methods. The resulting method accepts two variables but if shift is undefined by the user it will set the shift to the default value of 5.